Centralized props
