+++ src/index.js
@@ -1,7 +1,7 @@
 //@flow
 import * as React from 'react'
+import Wrap from './Wrap'
 
-import Wrap, { defaultProps } from './Wrap'
 // Stylized
 export { default as Facebook } from './stylized/FacebookStyle'
 export { default as Instagram } from './stylized/InstagramStyle'
@@ -21,13 +21,22 @@ export type Props = {
   className: string,
 }
 
+export const defaultProps = {
+  speed: 2,
+  width: 400,
+  height: 130,
+  primaryColor: '#f0f0f0',
+  secondaryColor: '#e0e0e0',
+  preserveAspectRatio: 'xMidYMid meet',
+}
+
+const InitialComponent = props => (
+  <rect x="0" y="0" rx="5" ry="5" width={props.width} height={props.height} />
+)
+
 const ContentLoader = (props: Props) => {
   const mergedProps = { ...defaultProps, ...props }
-  const children = props.children ? (
-    props.children
-  ) : (
-    <rect x="0" y="0" rx="5" ry="5" width={mergedProps.width} height={mergedProps.height} />
-  )
+  const children = props.children ? props.children : <InitialComponent {...mergedProps} />
 
   return <Wrap {...mergedProps}>{children}</Wrap>
 }
