+++ src/stylized/BulletListStyle.js
@@ -1,10 +1,10 @@
 //@flow
 import * as React from 'react'
-import Wrap, { defaultProps } from '../Wrap'
+import ContentLoader from '../index'
 import type { WrapProps } from '../Wrap'
 
 const BulletListStyle = (props: WrapProps): React.Element<*> => (
-  <Wrap {...defaultProps} {...props}>
+  <ContentLoader {...props}>
     <circle cx="10" cy="20" r="8" />
     <rect x="25" y="15" rx="5" ry="5" width="220" height="10" />
     <circle cx="10" cy="50" r="8" />
@@ -13,7 +13,7 @@ const BulletListStyle = (props: WrapProps): React.Element<*> => (
     <rect x="25" y="75" rx="5" ry="5" width="220" height="10" />
     <circle cx="10" cy="110" r="8" />
     <rect x="25" y="105" rx="5" ry="5" width="220" height="10" />
-  </Wrap>
+  </ContentLoader>
 )
 
 export default BulletListStyle
