+++ src/stylized/CodeStyle.js
@@ -1,10 +1,10 @@
 //@flow
 import * as React from 'react'
-import Wrap, { defaultProps } from '../Wrap'
+import ContentLoader from '../index'
 import type { WrapProps } from '../Wrap'
 
 const CodeStyle = (props: WrapProps): React.Element<*> => (
-  <Wrap {...defaultProps} {...props}>
+  <ContentLoader {...props}>
     <rect x="0" y="0" rx="3" ry="3" width="70" height="10" />
     <rect x="80" y="0" rx="3" ry="3" width="100" height="10" />
     <rect x="190" y="0" rx="3" ry="3" width="10" height="10" />
@@ -17,7 +17,7 @@ const CodeStyle = (props: WrapProps): React.Element<*> => (
     <rect x="185" y="40" rx="3" ry="3" width="60" height="10" />
 
     <rect x="0" y="60" rx="3" ry="3" width="30" height="10" />
-  </Wrap>
+  </ContentLoader>
 )
 
 export default CodeStyle
